requirements.txt
requests
python-telegram-bot==11.1.0
odata1cw
